deps/v8/test/inspector/debugger/wasm-stepping-in-from-js-expected.txt

Tests stepping from javascript into wasm

Running test: test
Calling instantiate function.
Waiting for wasm scripts to be parsed.
Ignoring script with url v8://test/instantiate
Got wasm script: wasm://wasm/fd1bd682
Setting breakpoint on i32.const
{
    columnNumber : 37
    lineNumber : 0
    scriptId : <scriptId>
}
paused
function test() {
  #debugger;
  instance.exports.main(1);

Debugger.stepInto
paused
  debugger;
  #instance.exports.main(1);
}

Debugger.stepInto
paused
Script wasm://wasm/fd1bd682 byte offset 35: Wasm opcode 0x20 (kExprLocalGet)
Debugger.resume
paused
Script wasm://wasm/fd1bd682 byte offset 37: Wasm opcode 0x41 (kExprI32Const)
Debugger.resume
exports.main returned!
